+++ packages/xchain-evm/src/providers/index.ts
@@ -1,2 +1,0 @@
-export * from './etherscan/types'
-export * from './etherscan/etherscan-data-provider'