+++ packages/xchain-evm/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.3.0 (2023-08-10)
+
+## Update
+
+- add support for fallback on providers
+- Update to use `xchain-evm-providers`
+
 # v0.2.2 (2023-05-18)
 
 ## Add